Extracting the class PartialFunction from brain_functools and inserting it into objects. Implementing its qname as the class's name. Testing this qname in LookupMixin._filter_stmts (node_classes)
